requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
pydantic==2.5.0
cachetools==5.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
PyGithub>=2.0.0